feat(linux/clipboard): support ext-data-control v1 alongside wlr
Add an ext-data-control-v1 backend so GNOME mutter 47+, KDE Plasma 6,
and recent wlroots compositors get the same native Wayland watcher and
read/write path that wlr-data-control already provides. Where both
protocols are advertised, ext is preferred (it's the standardized
cross-compositor protocol); UC_FORCE_DATA_CONTROL=ext|wlr lets local dev
override the choice for regression checks.

The two protocols are bit-identical in shape but use distinct concrete
types in wayland-rs's bindings, and Dispatch impls (plus
event_created_child!) cannot be made generic. Helpers stay shared via
an OfferLike trait (transfer / snapshot / write_payload), while each
protocol owns its complete EventLoop + Clipboard worker + Dispatch
impls under `wayland/protocol/{wlr,ext}.rs`. The self-echo deadlock
guard from the wlr worker is mirrored on the ext side.

Verified on niri (which advertises both): default selects ext;
UC_FORCE_DATA_CONTROL=wlr regresses cleanly. Both watcher and clipboard
test examples pass on either path.
