+++ src-tauri/crates/uc-platform/src/clipboard/platform/linux/wayland/mod.rs
@@ -1,34 +1,36 @@
 //! Native Wayland clipboard backend for Linux.
 //!
-//! ## Scope
+//! Two compositor protocols implement clipboard control: `wlr-data-control-v1`
+//! (niri / sway / hyprland / KDE / wlroots) and `ext-data-control-v1` (GNOME
+//! mutter ≥ 47, KDE Plasma 6, recent wlroots compositors). They are
+//! bit-identical in shape, so the structural code is shared via small traits
+//! ([`backend::OfferLike`]) and protocol-agnostic helpers
+//! ([`transfer`], [`snapshot`], [`write_payload`]). The per-protocol pieces
+//! that wayland-rs's `Dispatch` machinery requires to be concrete (event
+//! pattern matching, `event_created_child!` invocations, manager bindings)
+//! live in [`protocol::wlr`] and [`protocol::ext`].
 //!
-//! Phase 2a (this module's initial cut): event-driven clipboard *watcher* on
-//! the `wlr-data-control-unstable-v1` protocol, exposed via
-//! [`event_loop::WaylandEventLoop`]. This is the part that lights up the
-//! "compositor told me the clipboard changed" path and was the original
-//! motivation for the rewrite — Linux X11 clients (the previous default) are
-//! invisible to native Wayland app copies.
+//! Selection happens at construction time via [`protocol::try_new_event_loop`]
+//! / [`protocol::try_new_clipboard`]:
 //!
-//! Phase 2b: read/write `SystemClipboardPort` impl (`WaylandClipboard`).
-//! Phase 5 (follow-up): `ext-data-control-v1` (GNOME mutter 47+) on top of
-//! the same offer/source state machines.
+//! 1. `UC_FORCE_DATA_CONTROL=ext|wlr` overrides the default — useful for local
+//!    verification on compositors (like niri) that advertise both protocols.
+//! 2. Otherwise prefer `ext-data-control` (the standardized future-proof
+//!    choice), fall back to `wlr-data-control` (broader real-world coverage
+//!    today).
 //!
-//! ## Compositor support matrix (wlr-data-control v1)
-//!
-//! - niri ≥ 0.1, sway ≥ 1.6, Hyprland, KDE Plasma 5/6 — full support
-//! - wlroots-based compositors generally — full support
-//! - GNOME mutter — **not supported** (only ext-data-control v1, see Phase 5)
-//!
-//! On a session whose compositor doesn't advertise the manager global,
-//! [`event_loop::WaylandEventLoop::try_new`] returns `Ok(None)` and
-//! [`super::super::build_event_loop`] falls back to the legacy `clipboard_rs`
-//! adapter.
+//! On a session whose compositor advertises neither manager,
+//! [`event_loop::WaylandEventLoop::try_new`] / [`clipboard::WaylandClipboard::try_new`]
+//! return `Ok(None)` and [`super::super::build_event_loop`] / `LinuxClipboard::new`
+//! fall back to the legacy `clipboard_rs` adapter.
 
+mod backend;
 mod clipboard;
 mod event_loop;
+mod protocol;
 mod snapshot;
-mod state;
 mod transfer;
+mod write_payload;
 
 pub(super) use clipboard::WaylandClipboard;
 pub(super) use event_loop::WaylandEventLoop;
